Add handlers module
Move RunRotatingHandler, SlackHandler to handlers module
Remove handlers.py
