+++ delogger/modes/file.py
@@ -4,7 +4,7 @@ from pathlib import Path
 from typing import Optional
 
 from delogger.base import DeloggerBase
-from delogger.handlers import RunRotatingHandler
+from delogger.handlers.run_rotating import RunRotatingHandler
 from delogger.modes.base import ModeBase
